Все изменения по проекту в соответствии с todo отображени возле них самих

+++ application/controllers/controller_dove.php
@@ -33,14 +33,28 @@ class Controller_Dove extends Controller
             $dove_id = $_GET['id'];
             $dove = new Model_Dove();
             $dove_template = $dove->findOne(['id' => $dove_id]);
-            $this->view->generate(
+            if($dove_template === false) {
+                $response = "Данного голубя не существует";
+                $this->view->generate(
+                    'message_view.php',
+                    'template_view.php',
+                    ['message' => $response]
+                );
+            }
+            else
+                $this->view->generate(
                 'dove_view.php',
                 'template_view.php',
                 ['dove' => $dove_template]
-            );
+                );
         }
         else {
-            echo "Не вибран соответствующий продукт";
+            $response = "Не вибран соответствующий продукт";
+            $this->view->generate(
+                'message_view.php',
+                'template_view.php',
+                ['message' => $response]
+            );
         }
     }
 
@@ -49,18 +63,21 @@ class Controller_Dove extends Controller
         $dove ->findAll($_POST);
     }
 
-    public function action_add_record_form()//insert
+    public function action_add_dove_form()//insert
     {
         $dove = new Model_Dove();
         $htmlForm = $dove->generateForm();
-        $this->view->generate('addrecord_view.php', 'template_view.php', $htmlForm);
+        $this->view->generate('addrecord_view.php', 'template_view.php',
+            ['htmlForm' => $htmlForm, 'path' => '/dove/add_dove']);
     }
 
-    public function action_add_record()//insert
+    public function action_add_dove()//insert
     {
         $dove = new Model_Dove();// {title: null, price: null}
         $dove->load($_POST);// {title: qweqwe, price: 120}
         $dove->newSave();
+        $response = "Добавили голубя!";
+        $this->view->generate('message_view.php', 'template_view.php', ['message' => $response]);
     }
 
     public function action_update_dove_form()
@@ -68,17 +85,18 @@ class Controller_Dove extends Controller
         $dove = new Model_Dove();
         $dove_template = $dove->findOne(['id' => $_GET['id']]);
         $dove->map($dove_template);
+        //$htmlForm = "<input type='hidden' name='id' value='".$_GET['id']."'>";
         $htmlForm = $dove->generateForm(true);
         $this->view->generate('update_view.php',
-            'template_view.php', $htmlForm);
+            'template_view.php', ['htmlForm' => $htmlForm, 'path' => '/dove/update_dove']);
     }
 
     public function action_update_dove()
     {
         $dove = new Model_Dove();// {title: null, price: null}
-        //$dove->findOne(['id' => $_POST['id']]);// {title: qweqwe, price: 120}
         $dove->load($_POST);
         $dove->newSave();
+        header("Location: /dove/dove_show?id={$_POST['id']}");
     }
 
     public function action_delete_dove(){
@@ -86,11 +104,24 @@ class Controller_Dove extends Controller
             $dove_id = $_GET['id'];
             $dove = new Model_Dove();
             $dove_template = $dove->findOne(['id' => $dove_id]);
-            $dove->load($dove_template);
-            $dove->delete();
+            if($dove_template === false){
+                $response = "Вибран не существующий голубь";
+                $this->view->generate('message_view.php', 'template_view.php',
+                ['message' => $response]);
+            }
+            else {
+                $dove->load($dove_template);
+                $dove->delete();
+                $response = "Dovik удален!";
+                $this->view->generate('message_view.php', 'template_view.php',
+                    ['message' => $response]);
+                //header("Location: /dove/index");
+            }
         }
         else {
-            echo "Не вибран соответствующий продукт";
+            $response = "Не вибран соотвествующий голубок!";
+            $this->view->generate('message_view.php', 'template_view.php',
+                ['message' => $response]);
         }
     }
 }
